Uploading v0.62F
v0.62F (Final):
- Introduced MPI_int_buffer for sending an integer scalar - fixed up all
the MPI communication subroutine calls.
- Settled on without the -i8 flag - so all the MPI communication uses
the default MPI_INTEGER (which has KIND=4). I really don't need 64-bit
integers.
- Did I mention how much Fortran is stoooopid?
- Implemented recording of qubit selection phase with no premarture
termination.
- Implemented careful selection (using the ChooseN function) of N after
finding cases where the original code would accidentally bloat a matrix
with size exactly matching a power of 2.
- Okay this SHOULD be the final release version - pending approval by
results.
- Results approved - THIS IS THE FINAL RELEASE VERSION HUZZAH!

+++ OptQC_Bridge.f90
@@ -9,7 +9,7 @@ type(prog_args) :: args_obj
 character(len=128) :: fin
 integer :: fdum, fstat
 ! MPI variables
-integer(4) :: root, my_rank, p, ierr
+integer :: root, my_rank, p, ierr
 
 ! Initialize MPI environment
 call MPI_Init(ierr)
